refactor(predictions): rename icon const leaf and drop the barrel re-export
