+++ src/features/polymarket/components/league-icon/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_SIZE = 24;
